Rename list instances to get instances (#258)
Moving away from `kudo list instances` to more kubectl-like feel with `kudo get instances`
